+++ lyr-core/src/main/java/com/lyr/rule/RuleFactory.java
@@ -7,6 +7,7 @@ import com.lyr.rule.glue.ScanGlueSessionActiveWithLongIdleTimeoutRuleExecution;
 import com.lyr.rule.lambda.ScanLambdaFunctionWithDisallowedArchitectureRuleExecution;
 import com.lyr.rule.lambda.ScanLambdaFunctionWithUnboundedConcurrencyRuleExecution;
 import com.lyr.rule.lambda.ScanLambdaFunctionWithXrayTracingNotEnabledRuleExecution;
+import com.lyr.rule.lambda.ScanLambdaFunctionWithoutAnyActiveTriggerRuleExecution;
 import com.lyr.util.RuleDefinition;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -46,6 +47,11 @@ public class RuleFactory {
                         ruleDefinition,
                         RuleSetConfig.getRuleSetConfig().getRuleConfig(ruleDefinition),
                         new ScanLambdaFunctionWithXrayTracingNotEnabledRuleExecution());
+            case SCAN_LAMBDA_FUNCTION_WITHOUT_ANY_ACTIVE_TRIGGER ->
+                new Rule(
+                        ruleDefinition,
+                        RuleSetConfig.getRuleSetConfig().getRuleConfig(ruleDefinition),
+                        new ScanLambdaFunctionWithoutAnyActiveTriggerRuleExecution());
         };
     }
 }