+++ lyr-core/src/main/java/com/lyr/rule/lambda/ScanLambdaFunctionWithDisallowedArchitectureRuleExecution.java
@@ -5,7 +5,7 @@ import com.lyr.report.model.Finding;
 import com.lyr.rule.RuleExecutionStrategy;
 import com.lyr.rule.lambda.config.ScanLambdaFunctionWithDisallowedArchitectureRuleConfig;
 import com.lyr.services.lambda.LambdaConnector;
-import java.util.Locale;
+import com.lyr.util.StringUtil;
 import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.services.lambda.model.FunctionConfiguration;
 
@@ -17,9 +17,8 @@ public class ScanLambdaFunctionWithDisallowedArchitectureRuleExecution
     public ImmutableList<Finding> execute(final ScanLambdaFunctionWithDisallowedArchitectureRuleConfig config) {
         final var listOfLambdaFunctionConfigurations = LambdaConnector.create().listLambdaFunctionConfigurations();
         final var findings = listOfLambdaFunctionConfigurations.stream()
-                .filter(functionConfig -> functionConfig
-                        .architecturesAsStrings()
-                        .contains(config.getDisallowedArchitecture().toLowerCase(Locale.ROOT)))
+                .filter(functionConfig -> StringUtil.containsIgnoreCase(
+                        functionConfig.architecturesAsStrings(), config.getDisallowedArchitecture()))
                 .map(FunctionConfiguration::functionName)
                 .map(Finding::byId)
                 .collect(ImmutableList.toImmutableList());