+++ lyr-core/src/test/java/com/lyr/rule/lambda/config/ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfigTest.java
@@ -1,6 +1,6 @@
 package com.lyr.rule.lambda.config;
 
-import static com.lyr.rule.lambda.config.ScanLambdaFunctionWithDisallowedArchitectureRuleConfig.DISALLOWED_ARCHITECTURE_CONFIG_KEY;
+import static com.lyr.rule.lambda.config.ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfig.TRIGGER_STATES_CONSIDERED_AS_ACTIVE_KEY;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -11,13 +11,14 @@ import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.times;
 
 import com.lyr.exception.rule.config.BadRuleConfigException;
+import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.MockedStatic;
 
-class ScanLambdaFunctionWithDisallowedArchitectureRuleConfigTest {
+class ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfigTest {
 
     static MockedStatic<BadRuleConfigException> mockedBadRuleConfigExceptionStatic =
             mockStatic(BadRuleConfigException.class, CALLS_REAL_METHODS);
@@ -36,7 +37,7 @@ class ScanLambdaFunctionWithDisallowedArchitectureRuleConfigTest {
     void testThatRuleConfigIsCopiedCorrectly() {
         // Given
         final var expectedConfig =
-                ScanLambdaFunctionWithDisallowedArchitectureRuleConfig.builder().build();
+                ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfig.builder().build();
 
         // When
         final var actualCopiedConfig = expectedConfig.copy();
@@ -49,10 +50,10 @@ class ScanLambdaFunctionWithDisallowedArchitectureRuleConfigTest {
     @Test
     void testThatRuleConfigIsConvertedToAMapSuccessfully() {
         // Given
-        final var expectedConfig = ScanLambdaFunctionWithDisallowedArchitectureRuleConfig.builder()
-                .disallowedArchitecture("arm64")
+        final var expectedConfig = ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfig.builder()
+                .triggerStatesConsideredAsActive(List.of("a", "b"))
                 .build();
-        final var expectedConfigMap = Map.of(DISALLOWED_ARCHITECTURE_CONFIG_KEY, "arm64");
+        final var expectedConfigMap = Map.of(TRIGGER_STATES_CONSIDERED_AS_ACTIVE_KEY, "[\"a\", \"b\"]");
 
         // When
         final var actualConfigMap = expectedConfig.getConfigAsStringMap();
@@ -65,24 +66,23 @@ class ScanLambdaFunctionWithDisallowedArchitectureRuleConfigTest {
     void testThatRuleConfigValidationPassesGivenValidRuleConfig() {
         // Given
         final var ruleConfig =
-                ScanLambdaFunctionWithDisallowedArchitectureRuleConfig.builder().build();
+                ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfig.builder().build();
 
         // When & Then
         assertThatNoException().isThrownBy(ruleConfig::validate);
     }
 
     @Test
-    void testThatRuleConfigValidationFailsGivenRuleConfigWithUnsupportedArchitecture() {
+    void testThatRuleConfigValidationFailsGivenRuleConfigWithUnsupportedTriggerStates() {
         // Given
-        final ScanLambdaFunctionWithDisallowedArchitectureRuleConfig ruleConfig =
-                ScanLambdaFunctionWithDisallowedArchitectureRuleConfig.builder()
-                        .disallowedArchitecture("skynet")
-                        .build();
+        final var ruleConfig = ScanLambdaFunctionWithoutAnyActiveTriggerRuleConfig.builder()
+                .triggerStatesConsideredAsActive(List.of("a", "b"))
+                .build();
 
         // When & Then
         assertThatThrownBy(ruleConfig::validate).isInstanceOf(BadRuleConfigException.class);
 
         mockedBadRuleConfigExceptionStatic.verify(
-                () -> BadRuleConfigException.forUnsupportedValue(anyString(), anyString(), anyCollection()), times(1));
+                () -> BadRuleConfigException.forUnsupportedValues(anyString(), anyString(), anyCollection()), times(1));
     }
 }